# docs/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "oasis_sid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Dashboard and back-office API. Errors are application/problem+json; money is integer cents; instants are ISO-8601 UTC; business dates are YYYY-MM-DD in the business timezone.",
    "title": "Oasis Auto Spa API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/employees/{id}/deactivate": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "allowedCount": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "avatarColor": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "daysPerWeek": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "deactivatedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "effectivePermissions": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "key": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "limit": {
                            "anyOf": [
                              {
                                "maximum": 9007199254740991,
                                "minimum": -9007199254740991,
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "module": {
                            "type": "string"
                          },
                          "on": {
                            "type": "boolean"
                          },
                          "ov": {
                            "anyOf": [
                              {
                                "enum": [
                                  "allow",
                                  "deny"
                                ],
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "src": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "key",
                          "module",
                          "label",
                          "on",
                          "src",
                          "ov"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "email": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "employmentType": {
                      "enum": [
                        "full_time",
                        "part_time",
                        "contractor"
                      ],
                      "type": "string"
                    },
                    "exceptionCount": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "first": {
                      "type": "string"
                    },
                    "fullName": {
                      "type": "string"
                    },
                    "hasLogin": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string"
                    },
                    "initials": {
                      "type": "string"
                    },
                    "last": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "overrides": {
                      "additionalProperties": {
                        "enum": [
                          "allow",
                          "deny"
                        ],
                        "type": "string"
                      },
                      "propertyNames": {
                        "type": "string"
                      },
                      "type": "object"
                    },
                    "payType": {
                      "anyOf": [
                        {
                          "enum": [
                            "hourly",
                            "commission",
                            "salary"
                          ],
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "phone": {
                      "type": "string"
                    },
                    "phoneE164": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "rateText": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "roles": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "key": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "key",
                          "name"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "schedule": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "from": {
                            "type": "string"
                          },
                          "fromMin": {
                            "maximum": 9007199254740991,
                            "minimum": -9007199254740991,
                            "type": "integer"
                          },
                          "on": {
                            "type": "boolean"
                          },
                          "to": {
                            "type": "string"
                          },
                          "toMin": {
                            "maximum": 9007199254740991,
                            "minimum": -9007199254740991,
                            "type": "integer"
                          },
                          "weekday": {
                            "maximum": 9007199254740991,
                            "minimum": -9007199254740991,
                            "type": "integer"
                          }
                        },
                        "required": [
                          "weekday",
                          "on",
                          "fromMin",
                          "toMin",
                          "from",
                          "to"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "skills": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "status": {
                      "enum": [
                        "active",
                        "invited",
                        "inactive"
                      ],
                      "type": "string"
                    },
                    "statusLabel": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "version": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "first",
                    "last",
                    "name",
                    "fullName",
                    "initials",
                    "title",
                    "phone",
                    "phoneE164",
                    "email",
                    "status",
                    "statusLabel",
                    "employmentType",
                    "payType",
                    "rateText",
                    "skills",
                    "avatarColor",
                    "roles",
                    "exceptionCount",
                    "daysPerWeek",
                    "hasLogin",
                    "version",
                    "createdAt",
                    "deactivatedAt",
                    "schedule",
                    "overrides",
                    "effectivePermissions",
                    "allowedCount"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Deactivate (revokes their sessions; the last active Super Admin cannot be deactivated)",
        "tags": [
          "people"
        ],
        "x-oasis-access": "team.edit"
      }
    },
    "/api/v1/employees/{id}/effective-permissions": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "allowedCount": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "items": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "key": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "limit": {
                            "anyOf": [
                              {
                                "maximum": 9007199254740991,
                                "minimum": -9007199254740991,
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "module": {
                            "type": "string"
                          },
                          "on": {
                            "type": "boolean"
                          },
                          "ov": {
                            "anyOf": [
                              {
                                "enum": [
                                  "allow",
                                  "deny"
                                ],
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "src": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "key",
                          "module",
                          "label",
                          "on",
                          "src",
                          "ov"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "total": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "allowedCount",
                    "total"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Per-permission { on, src, ov, limit } for one employee",
        "tags": [
          "people"
        ],
        "x-oasis-access": "team.view"
      }
    },
    "/api/v1/employees/{id}/invite/resend": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "channel": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string"
                    },
                    "link": {
                      "type": "string"
                    },
                    "sent": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "sent",
                    "channel",
                    "expiresAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Issue a fresh 7-day invite and send it",
        "tags": [
          "people"
        ],
        "x-oasis-access": "team.edit"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ]
}